+++ pixano/data/dataset_info.py
@@ -25,8 +25,8 @@ class DatasetInfo(BaseModel):
         id (str): Dataset ID
         name (str): Dataset name
         description (str): Dataset description
-        estimated_size (str): Dataset estimated size
-        num_elements (int): Number of elements in dataset
+        estimated_size (str, optional): Dataset estimated size
+        num_elements (int, optional): Number of elements in dataset
         preview (str, optional): Dataset preview
         splits (list[str]): Dataset splits
         tables (dict[str, list], optional): Dataset tables
@@ -43,8 +43,63 @@ class DatasetInfo(BaseModel):
     tables: Optional[dict[str, list]]
     categories: Optional[list[dict]]
 
+    def __init__(
+        self,
+        id: str,
+        name: str,
+        description: str,
+        estimated_size: str = None,
+        num_elements: int = None,
+        preview: str = None,
+        splits: list[str] = None,
+        tables: dict[str, list] = None,
+        categories: list[dict] = None,
+    ):
+        """Initialize Bounding box
+
+        Args:
+            id (str): Dataset ID
+            name (str): Dataset name
+            description (str): Dataset description
+            estimated_size (str, optional): Dataset estimated size. Defaults to None.
+            num_elements (int, optional): Number of elements in dataset. Defaults to None.
+            preview (str, optional): Dataset preview. Defaults to None.
+            splits (list[str]): Dataset splits. Defaults to None.
+            tables (dict[str, list], optional): Dataset tables. Defaults to None.
+            categories (list[dict], optional): Dataset categories. Defaults to None.
+        """
+
+        # Define public attributes through Pydantic BaseModel
+        super().__init__(
+            id=id,
+            name=name,
+            description=description,
+            estimated_size=estimated_size,
+            num_elements=num_elements,
+            preview=preview,
+            splits=splits,
+            tables=tables,
+            categories=categories,
+        )
+
     def save(self, save_dir: Path):
         """Save DatasetInfo to json file"""
 
         with open(save_dir / "db.json", "w", encoding="utf-8") as f:
-            json.dump(self.dict(), f)
+            json.dump(self.model_dump(), f)
+
+    @staticmethod
+    def from_json(json_fp: Path) -> "DatasetInfo":
+        """Read DatasetInfo from JSON file
+
+        Args:
+            json_fp (Path): JSON file path
+
+        Returns:
+            DatasetInfo: DatasetInfo
+        """
+
+        with open(json_fp) as json_file:
+            info_json = json.load(json_file)
+
+        return DatasetInfo.model_validate(info_json)