+++ pixano/data/importers/__init__.py
@@ -12,7 +12,7 @@
 # http://www.cecill.info
 
 from pixano.data.importers.coco_importer import COCOImporter
-from pixano.data.importers.dota_importer_test import DOTAImporter
+from pixano.data.importers.dota_importer import DOTAImporter
 from pixano.data.importers.image_importer import ImageImporter
 from pixano.data.importers.importer import Importer