+++ pixano/data/importers/image_importer.py
@@ -27,13 +27,20 @@ class ImageImporter(Importer):
         info (DatasetInfo): Dataset information
     """
 
-    def __init__(self, name: str, description: str, splits: list[str] = ["dataset"]):
+    def __init__(
+        self,
+        name: str,
+        description: str,
+        splits: list[str] = None,
+        media_fields: dict[str, str] = {"image": "image"},
+    ):
         """Initialize Image Importer
 
         Args:
             name (str): Dataset name
             description (str): Dataset description
-            splits (list[str], optional): Dataset splits. Defaults to ["dataset"] for datasets with no subfolders for splits.
+            splits (list[str], optional): Dataset splits. Defaults to None for datasets with no subfolders for splits.
+            media_fields (dict[str, str]): Dataset media fields, with field names as keys and field types as values. Default to {"image": "image"}.
         """
 
         tables = {
@@ -47,16 +54,32 @@ class ImageImporter(Importer):
                     },
                 }
             ],
-            "media": [
-                {
-                    "name": "image",
-                    "fields": {
-                        "id": "str",
-                        "image": "image",
-                    },
-                }
-            ],
+            "media": [],
         }
+
+        # Add media fields to tables
+        for field_name, field_type in media_fields.items():
+            table_exists = False
+            # If table for given field type exists
+            for media_table in tables["media"]:
+                if field_type == media_table["name"] and not table_exists:
+                    media_table["fields"][field_name] = field_type
+                    table_exists = True
+            # Else, create that table
+            if not table_exists:
+                tables["media"].append(
+                    {
+                        "name": field_type,
+                        "fields": {
+                            "id": "str",
+                            field_name: field_type,
+                        },
+                    }
+                )
+
+        # If no splits given, define a default single dataset split
+        if not splits:
+            splits = ["dataset"]
 
         # Initialize Importer
         super().__init__(name, description, tables, splits)