+++ pixano/data/importers/importer.py
@@ -83,10 +83,9 @@ class Importer(ABC):
         self.info.num_elements = len(ds_tables["main"]["db"])
         self.info.estimated_size = estimate_size(import_dir)
 
+        # Save DatasetInfo
         with tqdm(desc="Creating dataset info file", total=1) as progress:
-            # Create spec.json
-            with open(import_dir / "db.json", "w", encoding="utf-8") as f:
-                json.dump(self.info.dict(), f)
+            self.info.save(import_dir)
             progress.update(1)
 
     def create_preview(
@@ -232,6 +231,12 @@ class Importer(ABC):
         for table_group, tables in self.info.tables.items():
             for table in tables:
                 ds_tables[table_group][table["name"]] = ds.open_table(table["name"])
+
+        # Raise error if generated dataset is empty
+        if len(ds_tables["main"]["db"]) == 0:
+            raise FileNotFoundError(
+                "Generated dataset is empty. Please make sure that the paths to your media files are correct, and that they each contain subfolders for your splits."
+            )
 
         # Copy media directories if portable
         if portable and "media" in ds_tables: