+++ pixano/utils/boxes.py
@@ -16,25 +16,26 @@ import numpy as np
 from pixano.utils.image import rle_to_mask, urle_to_rle
 
 
-def denormalize_coords(coord: list[float], height: int, width: int) -> list[float]:
+def denormalize_coords(coord: list[float], height: int, width: int, rounded_int=True) -> list[float]:
     """Denormalize coordinates
 
     Args:
         coord (list[float]): Normalized coordinates
         height (int): Height
         width (int): Width
+        rounded_int (bool): True to round denormalized float to nearest integer. Default to True
 
     Returns:
-        list[float]: Unnormalized coordinates
+        list[float]: Unnormalized coordinates,
     """
 
     denorm = []
 
     for i, c in enumerate(coord):
         if i % 2 == 0:
-            denorm.append(c * width)
+            denorm.append(round(c * width) if rounded_int else c * width)
         else:
-            denorm.append(c * height)
+            denorm.append(round(c * height) if rounded_int else c * height)
 
     return denorm