+++ tests/data/importers/test_dota_importer.py
@@ -17,26 +17,26 @@ from pathlib import Path
 
 import pyarrow as pa
 
-from pixano.core import BBoxType, CompressedRLEType, ImageType
-from pixano.data.importers.coco_importer import COCOImporter
+from pixano.core import BBoxType, ImageType
+from pixano.data import DOTAImporter
 
 
-class COCOImporterTestCase(unittest.TestCase):
+class DOTAImporterTestCase(unittest.TestCase):
     def setUp(self):
         self.input_dirs = {
-            "image": Path("unit_testing/assets/coco_dataset/image"),
-            "objects": Path("unit_testing/assets/coco_dataset"),
+            "image": Path("tests/assets/coco_dataset/image"),
+            "objects": Path("tests/assets/coco_dataset/objects"),
         }
-        self.importer = COCOImporter(
-            name="COCO",
-            description="COCO dataset",
+        self.importer = DOTAImporter(
+            name="DOTA",
+            description="DOTA dataset from COCO",
             splits=["val"],
         )
 
     def test_import_dataset(self):
         with tempfile.TemporaryDirectory() as temp_dir:
             # Set import directory
-            import_dir = Path(temp_dir) / "coco"
+            import_dir = Path(temp_dir) / "dota"
 
             # Import dataset
             dataset = self.importer.import_dataset(
@@ -50,8 +50,8 @@ class COCOImporterTestCase(unittest.TestCase):
             self.assertTrue(spec_json_path.exists())
 
             # Check db.json content
-            self.assertEqual("COCO", dataset.info.name)
-            self.assertEqual(3, dataset.info.num_elements)
+            self.assertEqual("DOTA", dataset.info.name)
+            self.assertEqual(dataset.info.num_elements, 1)
 
             # Check that db.lance exists
             db_lance_path = import_dir / "db.lance"
@@ -60,7 +60,7 @@ class COCOImporterTestCase(unittest.TestCase):
             # Check db.lance content
             ds = dataset.connect()
             table = ds.open_table("db")
-            self.assertEqual(len(table), 3)
+            self.assertEqual(len(table), 1)
             self.assertIn(pa.field("id", pa.string()), table.schema)
             self.assertIn(pa.field("split", pa.string()), table.schema)
 
@@ -71,7 +71,7 @@ class COCOImporterTestCase(unittest.TestCase):
             # Check image.lance content
             ds = dataset.connect()
             table = ds.open_table("image")
-            self.assertEqual(len(table), 3)
+            self.assertEqual(len(table), 1)
             self.assertIn(pa.field("id", pa.string()), table.schema)
             self.assertIn(pa.field("image", ImageType), table.schema)
 
@@ -82,7 +82,6 @@ class COCOImporterTestCase(unittest.TestCase):
             # Check objects.lance content
             ds = dataset.connect()
             table = ds.open_table("objects")
-            self.assertEqual(len(table), 39)
+            self.assertEqual(len(table), 323)
             self.assertIn(pa.field("id", pa.string()), table.schema)
             self.assertIn(pa.field("bbox", BBoxType), table.schema)
-            self.assertIn(pa.field("mask", CompressedRLEType), table.schema)